Share sliver agglomeration and short-edge collapse between cut paths
The hex and generic paths had drifted into near-verbatim copies of both
post-passes: collapse_short_edges was 258 lines of which only 21
differed, and those differences were entirely in the signature and in
how two inputs were derived. Moves intern, agglomerate and
collapse_short_edges into cut::cleanup and routes both callers through
them.

Only the two hex-shaped inputs needed lifting to the call sites:

- scales, the per-cell length yardstick: hex derives it from its source
  hexes and EDGES, the generic path from each source cell's own edges.
- whole, the anchored node set: hex flattens its retained-hex array, the
  generic path collects nodes of cells emitted with fraction 1.0.

collapse_short_edges also now takes the sign map directly rather than
either Tables type, which is all it ever wanted from them.

Verified the hex path is unchanged rather than merely still valid: a
fingerprint of cut() on bone at scales 3, 4 and 8 - hex count, polyhedral
count, node count, total signed volume and summed coordinate norms - is
identical to twelve digits against a worktree at the parent commit.

Net 671 lines removed with no behaviour change on either path.
